requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
boto3==1.17.46
botocore==1.20.46
Django==3.1.3
django-s3-folder-storage==0.5
django-storages==1.11.1
gunicorn==20.1.0
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.4
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
